Move the generate salt method into RandomDataGenerator
Keep all random functions in one class so it can be mocked
