+++ src/us/kbase/test/auth2/cryptutils/CryptUtilsTest.java
@@ -54,14 +54,6 @@ public class CryptUtilsTest {
 	}
 	
 	@Test
-	public void generateSalt() throws Exception {
-		// not much to test here other than it returns an 8 byte array
-		// even all 0s is a valid output
-		final byte[] salt = new PasswordCrypt().generateSalt();
-		assertThat("incorrect salt length", salt.length, is(8));
-	}
-	
-	@Test
 	public void shortSalt() throws Exception {
 		try {
 			new PasswordCrypt().getEncryptedPassword("f".toCharArray(), new byte[0]);
